fix(admin): modules should be resolved with module paths not system paths
